feat: +1846 matches (33% → 42%) — SDA, schedule-insns, DOL scan, template blast

+++ src/wip/version_diff/match_0x8002C32C_EIWallPart__GetVisible.cpp
@@ -1,16 +1,17 @@
 // EIWallPart::GetVisible(int)
 // Address: 0x8002C32C | Size: 48 bytes
+// FLAGS: -fno-elide-constructors
 
 struct EIWallPart {
     char pad[0x5C];
-    unsigned int m_visibleFlags;  // word access, not byte
+    unsigned int m_visibleFlags;
 };
 
 int EIWallPart__GetVisible(EIWallPart* self, int which) {
     if (which == 0) {
-        return (self->m_visibleFlags >> 7) & 1;
+        return (int)((self->m_visibleFlags >> 7) & 1U);
     } else if (which == 1) {
-        return self->m_visibleFlags & 1;
+        return (int)(self->m_visibleFlags & 1U);
     }
     return 0;
 }
